backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
torch
torchvision
opencv-python
pydantic
starlette
pytorch-lightning
torchmetrics
python-jose[cryptography]
passlib[bcrypt]
requests
python-multipart
openai
aiofiles
sqlalchemy

